src/requirements.txt
snowflake-snowpark-python>=1.33.0
snowflake-cli>=3.10.0
click>=8.1.8,<8.2.0
rich>=14.0.0,<14.1.0
pydantic>=2.11.7
pyyaml>=6.0.2
typing-extensions>=4.14.0,<4.14.2
python-dotenv>=1.1.1